requirements.txt
openai==0.28
django
python-dotenv
